apps/api/pyproject.toml
[project]
name = "translator-api"
version = "1.0.0"
description = "Translation + file conversion API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic-settings>=2.7",
    "python-multipart>=0.0.18",
    "openai>=1.60",
    "mistralai>=1.12",
    "chardet>=5.2",
    "python-docx>=1.1",
    "weasyprint>=62.0",
    "markdown-it-py>=3.0",
    "httpx>=0.28",
    "structlog>=24.0",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
whisper = ["faster-whisper>=1.1"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
